+++ pkg/telemetry/metrics/options.go
@@ -1,4 +1,4 @@
-package prometheus
+package metrics
 
 import (
 	"github.com/prometheus/client_golang/prometheus/collectors"
@@ -6,10 +6,11 @@ import (
 
 var defaultBuckets = []float64{0.005, 0.01, 0.025, 0.05, 0.1, 0.25, 0.5, 1, 2.5, 5, 10}
 
-// Option opts for monitor prometheus
+// Option defines the configuration options for the metrics tracer.
 type Option interface {
 	apply(cfg *promConfig)
 }
+
 type option func(cfg *promConfig)
 
 func (fn option) apply(cfg *promConfig) {